+++ core/src/main/java/com/datastrato/gravitino/storage/relational/JDBCBackend.java
@@ -37,9 +37,20 @@ import com.datastrato.gravitino.storage.relational.service.TableMetaService;
 import com.datastrato.gravitino.storage.relational.service.TopicMetaService;
 import com.datastrato.gravitino.storage.relational.service.UserMetaService;
 import com.datastrato.gravitino.storage.relational.session.SqlSessionFactoryHelper;
+import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.Statement;
 import java.util.List;
 import java.util.function.Function;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * {@link JDBCBackend} is a jdbc implementation of {@link RelationalBackend} interface. You can use
@@ -49,9 +60,13 @@ import java.util.function.Function;
  */
 public class JDBCBackend implements RelationalBackend {
 
+  private static final Logger LOG = LoggerFactory.getLogger(JDBCBackend.class);
+
   /** Initialize the jdbc backend instance. */
   @Override
   public void initialize(Config config) {
+    initJDBCBackend(config);
+
     SqlSessionFactoryHelper.getInstance().init(config);
     SQLExceptionConverterFactory.initConverter(config);
   }
@@ -277,5 +292,123 @@ public class JDBCBackend implements RelationalBackend {
   @Override
   public void close() throws IOException {
     SqlSessionFactoryHelper.getInstance().close();
+  }
+
+  enum EmbeddedJDBCBackendType {
+    H2("h2"),
+    DERBY("derby"),
+    SQLITE("sqlite"),
+    MYSQL("mysql");
+
+    private final String name;
+
+    EmbeddedJDBCBackendType(String name) {
+      this.name = name;
+    }
+
+    public static EmbeddedJDBCBackendType fromName(String name) {
+      for (EmbeddedJDBCBackendType backend : EmbeddedJDBCBackendType.values()) {
+        if (backend.name.equalsIgnoreCase(name)) {
+          return backend;
+        }
+      }
+
+      throw new IllegalArgumentException("Unknown EmbeddedJDBCBackend: " + name);
+    }
+  }
+
+  private static void initJDBCBackend(Config config) {
+    String embeddedJDBCUserName = "gravitino";
+    String embeddedJDBCPassword = "gravitino";
+    String storagePath = getStoragePath(config);
+
+    String jdbcUrl = config.get(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_URL);
+    String jdbcType = getJDBCType(jdbcUrl);
+
+    EmbeddedJDBCBackendType embeddedJDBCBackend = EmbeddedJDBCBackendType.fromName(jdbcType);
+
+    switch (embeddedJDBCBackend) {
+      case H2:
+        initH2Backend(config, embeddedJDBCUserName, embeddedJDBCPassword, storagePath);
+        break;
+      case DERBY:
+      case SQLITE:
+        throw new UnsupportedOperationException(embeddedJDBCPassword + " is not supported yet.");
+      case MYSQL:
+        break;
+      default:
+        throw new IllegalArgumentException("Unknown EmbeddedJDBCBackend: " + embeddedJDBCBackend);
+    }
+  }
+
+  private static String getJDBCType(String jdbcURI) {
+    if (jdbcURI.startsWith("jdbc:h2")) {
+      return "h2";
+    } else if (jdbcURI.startsWith("jdbc:derby")) {
+      return "derby";
+    } else if (jdbcURI.startsWith("jdbc:sqlite")) {
+      return "sqlite";
+    } else if (jdbcURI.startsWith("jdbc:mysql")) {
+      return "mysql";
+    } else {
+      throw new IllegalArgumentException("Unknown JDBC URI: " + jdbcURI);
+    }
+  }
+
+  private static void initH2Backend(
+      Config config, String h2UserName, String h2UserPassword, String storagePath) {
+    String gravitinoHome = System.getenv("GRAVITINO_HOME");
+    String originalJDBCUrl = config.get(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_URL);
+
+    String connectionUrl = constructH2URI(originalJDBCUrl, storagePath);
+    try (Connection connection =
+            DriverManager.getConnection(connectionUrl, h2UserName, h2UserPassword);
+        Statement statement = connection.createStatement()) {
+      String sqlContent =
+          FileUtils.readFileToString(
+              new File(gravitinoHome + "/scripts/h2/schema-h2.sql"), StandardCharsets.UTF_8);
+
+      statement.execute(sqlContent);
+    } catch (Exception e) {
+      LOG.error("Failed to create table for H2 database.", e);
+      throw new RuntimeException("Failed to create table for H2 database.", e);
+    }
+
+    config.set(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_URL, connectionUrl);
+    config.set(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_USER, h2UserName);
+    config.set(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_PASSWORD, h2UserPassword);
+    config.set(Configs.ENTITY_RELATIONAL_JDBC_BACKEND_DRIVER, "org.h2.Driver");
+  }
+
+  private static String constructH2URI(String originURI, String storagePath) {
+    if (!originURI.contains(":file:")) {
+      originURI = "jdbc:h2:file:" + storagePath;
+    }
+
+    if (!originURI.contains("DB_CLOSE_DELAY")) {
+      originURI = originURI + ";DB_CLOSE_DELAY=-1";
+    }
+
+    if (!originURI.contains("MODE")) {
+      originURI = originURI + ";MODE=MYSQL";
+    }
+
+    return originURI;
+  }
+
+  static String getStoragePath(Config config) {
+    String dbPath = config.get(Configs.ENTRY_RELATIONAL_JDBC_BACKEND_PATH);
+    if (StringUtils.isBlank(dbPath)) {
+      return Configs.DEFAULT_RELATIONAL_JDBC_BACKEND_PATH;
+    }
+
+    Path path = Paths.get(dbPath);
+    // Relative Path
+    if (!path.isAbsolute()) {
+      path = Paths.get(System.getenv("GRAVITINO_HOME"), dbPath);
+      return path.toString();
+    }
+
+    return dbPath;
   }
 }
