+++ core/src/test/java/com/datastrato/gravitino/storage/TestEntityStorage.java
@@ -6,7 +6,7 @@
 package com.datastrato.gravitino.storage;
 
 import static com.datastrato.gravitino.Configs.DEFAULT_ENTITY_KV_STORE;
-import static com.datastrato.gravitino.Configs.EMBEDDED_ENTITY_RELATIONAL_STORE;
+import static com.datastrato.gravitino.Configs.DEFAULT_ENTITY_RELATIONAL_STORE;
 import static com.datastrato.gravitino.Configs.ENTITY_KV_STORE;
 import static com.datastrato.gravitino.Configs.ENTITY_RELATIONAL_JDBC_BACKEND_DRIVER;
 import static com.datastrato.gravitino.Configs.ENTITY_RELATIONAL_JDBC_BACKEND_PASSWORD;
@@ -16,7 +16,6 @@ import static com.datastrato.gravitino.Configs.ENTITY_RELATIONAL_STORE;
 import static com.datastrato.gravitino.Configs.ENTITY_STORE;
 import static com.datastrato.gravitino.Configs.ENTRY_KV_ROCKSDB_BACKEND_PATH;
 import static com.datastrato.gravitino.Configs.ENTRY_RELATIONAL_JDBC_BACKEND_PATH;
-import static com.datastrato.gravitino.Configs.ENTRY_RELATIONAL_STORE_EMBEDDED_IMPLEMENTATION;
 import static com.datastrato.gravitino.Configs.RELATIONAL_ENTITY_STORE;
 import static com.datastrato.gravitino.Configs.STORE_DELETE_AFTER_TIME;
 import static com.datastrato.gravitino.Configs.STORE_TRANSACTION_MAX_SKEW_TIME;
@@ -112,10 +111,8 @@ public class TestEntityStorage {
       }
       dir.mkdirs();
       Mockito.when(config.get(ENTITY_STORE)).thenReturn(RELATIONAL_ENTITY_STORE);
-      Mockito.when(config.get(ENTITY_RELATIONAL_STORE))
-          .thenReturn(EMBEDDED_ENTITY_RELATIONAL_STORE);
+      Mockito.when(config.get(ENTITY_RELATIONAL_STORE)).thenReturn(DEFAULT_ENTITY_RELATIONAL_STORE);
       Mockito.when(config.get(ENTRY_RELATIONAL_JDBC_BACKEND_PATH)).thenReturn(DB_DIR);
-      Mockito.when(config.get(ENTRY_RELATIONAL_STORE_EMBEDDED_IMPLEMENTATION)).thenReturn("h2");
 
       // The following properties are used to create the JDBC connection; they are just for test, in
       // the real world,
